+++ studio/components/interfaces/AuditLogs/LogDetailsPanel.tsx
