+++ studio/components/interfaces/AuditLogs/index.ts
@@ -1,0 +1,4 @@
+import FilterPopover from './FilterPopover'
+import LogDetailsPanel from './LogDetailsPanel'
+
+export { FilterPopover, LogDetailsPanel }
